Spring Cursinho básico
Primeiro projeto de spring completinho com segurança implementado e recursos chaves

+++ AprendendoSozinho/Aulas/projetoSpring/src/main/java/com/example/projetoSpring/service/ProdutoService.java
@@ -1,5 +1,6 @@
 package com.example.projetoSpring.service;
 
+import com.example.projetoSpring.exception.RecursoNaoEncontradoException;
 import com.example.projetoSpring.model.Produto;
 import com.example.projetoSpring.repository.ProdutoRepository;
 import org.springframework.stereotype.Service;
@@ -19,8 +20,9 @@ public class ProdutoService {
         return produtoRepository.findAll();
     }
 
-    public Optional<Produto> buscarPorId(Long id){
-        return produtoRepository.findById(id);
+    public Produto buscarPorId(Long id){
+        return produtoRepository.findById(id)
+                .orElseThrow(() -> new RecursoNaoEncontradoException("Produto com o ID: "+id+" não encontrado"));
     }
 
     public Produto salvarProduto(Produto produto){
@@ -28,6 +30,9 @@ public class ProdutoService {
     }
 
     public void deletarProduto(Long id){
+        if(!produtoRepository.existsById(id)){
+            throw new RecursoNaoEncontradoException("Produto com o ID: "+id+" não encontrado");
+        }
         produtoRepository.deleteById(id);
     }
 }
